test(logger): move allocation assertions off the -race path
testing.AllocsPerRun counts the race detector's own allocations, so the
assertions failed in every CI job: all of them pass race=true. Under -race
each field measured one allocation more than its ceiling, and the bare
record measured one where it should have measured none.

Put them in //go:build !race files and add a test-alloc target, run from
the existing race-free bench job. -cover is unaffected and stays.
