+++ opensearchtransport/logger_debug_benchmark_internal_test.go
@@ -81,23 +81,6 @@ func eventFields() []eventField {
 	}
 }
 
-// TestTextDebugLoggerAllocations fails when a field method allocates more than
-// [eventFields] allows. The benchmarks report the same counts, but only for whoever
-// runs them; this is what makes a regression break the build.
-func TestTextDebugLoggerAllocations(t *testing.T) {
-	dl := &textDebugLogger{Output: io.Discard}
-
-	for _, f := range eventFields() {
-		t.Run(f.name, func(t *testing.T) {
-			// AllocsPerRun warms up f once before measuring, so the pooled event
-			// and its buffer are already in hand by the first counted run.
-			if got := int(testing.AllocsPerRun(100, func() { f.emit(dl) })); got > f.maxAllocs {
-				t.Errorf("allocations = %d, want at most %d", got, f.maxAllocs)
-			}
-		})
-	}
-}
-
 // BenchmarkTextDebugLoggerFields measures each field method against the built-in
 // logger, the one OPENSEARCH_GO_DEBUG and Config.EnableDebugLogger install. It is
 // the baseline an adapter has to be worth switching away from.
